Add consult acknowledgement apis

+++ Web.Data/Models/ConsultAcknowledgment.cs
@@ -8,7 +8,7 @@ namespace Web.Data.Models
     public partial class ConsultAcknowledgment
     {
         public int ConsultAcknowledgmentId { get; set; }
-        public int ConsultIdFk { get; set; }
+        public long ConsultIdFk { get; set; }
         public int UserIdFk { get; set; }
         public bool IsAcknowledge { get; set; }
         public int CreatedBy { get; set; }
